+++ src/main/java/org/fsg1/fmms/backend/endpoints/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains all the endpoints for the REST API.
+ */
+package org.fsg1.fmms.backend.endpoints;
