--- src/main/java/org/fsg1/fmms/backend/resources/package-info.java
@@ -1,4 +1,0 @@
-/**
- * This package contains all the resources for the REST API.
- */
-package org.fsg1.fmms.backend.resources;
