+++ src/main/java/org/fsg1/fmms/backend/services/package-info.java
@@ -1,5 +1,5 @@
 /**
- * This package contains all the service classes. These are the classes that the resources use to communicate
+ * This package contains all the service classes. These are the classes that the endpoints use to communicate
  * with the database.
  */
 package org.fsg1.fmms.backend.services;
